+++ src/lib/auth.ts
@@ -1,0 +1,20 @@
+import GitHubProvider from "next-auth/providers/github";
+import Kakao from "next-auth/providers/kakao";
+import Naver from "next-auth/providers/naver";
+
+export const authOptions = {
+  providers: [
+    GitHubProvider({
+      clientId: process.env.AUTH_GITHUB_ID!,
+      clientSecret: process.env.AUTH_GITHUB_SECRET!,
+    }),
+    Kakao({
+      clientId: process.env.AUTH_KAKAO_ID!,
+      clientSecret: process.env.AUTH_KAKAO_SECRET!,
+    }),
+    Naver({
+      clientId: process.env.AUTH_NAVER_ID!,
+      clientSecret: process.env.AUTH_NAVER_SECRET!,
+    }),
+  ],
+};
